refactor: package structure 정리
