+++ app/src/main/java/com/saehyun/talktopi_android/feature/intro/IntroActivity.kt
@@ -1,9 +1,12 @@
-package com.saehyun.talktopi_android
+package com.saehyun.talktopi_android.feature.intro
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import com.saehyun.talktopi_android.R
+import com.saehyun.talktopi_android.feature.sign_in.SignInActivity
+import com.saehyun.talktopi_android.feature.sign_up.SignUpActivity
 
 class IntroActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
