+++ app/src/main/java/com/saehyun/talktopi_android/feature/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.saehyun.talktopi_android
+package com.saehyun.talktopi_android.feature.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.saehyun.talktopi_android.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
