+++ app/src/main/java/com/saehyun/talktopi_android/feature/sign_in/SignInActivity.kt
@@ -1,10 +1,11 @@
-package com.saehyun.talktopi_android
+package com.saehyun.talktopi_android.feature.sign_in
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
-import android.widget.Button
+import com.saehyun.talktopi_android.feature.main.MainActivity
+import com.saehyun.talktopi_android.R
 
 class  SignInActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
