+++ app/src/main/java/com/saehyun/talktopi_android/feature/sign_up/SignUpAccountFragment.kt
@@ -1,10 +1,11 @@
-package com.saehyun.talktopi_android
+package com.saehyun.talktopi_android.feature.sign_up
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.saehyun.talktopi_android.R
 
 class SignUpAccountFragment : Fragment() {
