+++ app/src/main/java/com/saehyun/talktopi_android/feature/sign_up/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package com.saehyun.talktopi_android
+package com.saehyun.talktopi_android.feature.sign_up
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,6 +6,8 @@ import android.widget.Button
 import android.widget.FrameLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.saehyun.talktopi_android.feature.main.MainActivity
+import com.saehyun.talktopi_android.R
 
 enum class SignUpStep {
     ACCOUNT, PASSWORD
