+++ app/src/main/res/layout/activity_intro.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@drawable/bg_splash_image"
     android:layout_height="match_parent"
-    tools:context=".IntroActivity">
+    tools:context=".feature.intro.IntroActivity">
 
     <View
         android:layout_width="match_parent"
